notif_admin_panel/requirements.txt
django==3.1.12
python-dotenv==1.0.0
pytz
gunicorn==20.1.0
django-better-admin-arrayfield==1.4.2
psycopg2-binary==2.9.5